use code generator for SQS APIs
Using the code generator (ref: #96, #97) to generate SQS APIs.

There are some (minor) changes in the type names because of the way
the code generator generates names is different from what was written
earlier.

I have not attempted to map the earlier type and API names to
the new names. This will be a breaking change for those APIs.
Unless absolutely required, I plan to leave it this way and tag a new
major version when all APIs are migrated to this form.