Fix SQS tests assuming the query service response structure (#684)
* Fix SQS tests assuming the query service response structure

In PR 637, SQS changed from a `QueryService` to a `JSONService`. The
response obtained when interacting with SQS using the query protocol
includes a field called `<ActionName>Result`, where `<ActionName>` is
the action requested. The JSON protocol does not include that field.
Thus we can fix issue 683 by simply deleting that additional `getindex`
call.

* Also remove `ReceiveMessageResult`

* Helps if you read the docs

* Once again, the docs help

* ...And again, docs