Avoid duplicate `include`s causing warnings in tests
Currently we're calling `include` in a loop over supported backends. The
files being `include`d contain definitions which get overwritten when
`include` is called a second time, resulting in a lot of warnings that
make the test logs noisy. The approach taken here to address this is to
wrap things in expressions, define an anonymous module inside of the
loop over backends, `Core.eval` the expressions into the anonymous
module, and `Core.include` the files into it. There are better ways to
do this but this *should* get the job done with minimal required
changes.